Faster reading of training set images.
Determine ahead of time how many training examples will be read.
Preallocate X_train and y_train to be the size needed.
Fill in a slice of both for each image read.
For 200 images it was about 6 times faster, but yesterday’s approach (appending to numpy array) was O(n^2) so that could be as big an improvement as you want.